[mw/yarp] Fix the orientation of image
While here:
- move the Publisher in a specific file and adapt builder data file
- remove now incorrect patch
- improve python style

Moreover, it must improve the performance of the Yarp Image Publisher as
it avoids one useless copy
